requirements.txt
# Latest versions with Pydantic v2 support
openai>=1.50.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0

# Pydantic v2
pydantic>=2.0.0

# Vector store
chromadb>=0.5.0
tiktoken>=0.7.0